Prevent adding Policy that is not offered
Issue:
Adding a Policy to the client with a policy type id that does not exist breaks the application. Should not be able to add a policy to a client that does not exist inside InsuraBook. Should not be able to set an expiration date before today's date.

Fixes:
Throws policy type not found if attempting to add a policy to a client when the policy type does not exist in the application.
Throws an exception if provided expiration date of policy is before todays date.